Correct word size.

+++ src/include/riscostypes.h
@@ -10,8 +10,10 @@
 #ifndef _RISCOSE_TYPES_H
 #define _RISCOSE_TYPES_H
 
-typedef signed long int WORD_signed;
-typedef unsigned long int WORD;
+#include <stdint.h>
+
+typedef int32_t  WORD_signed;
+typedef uint32_t WORD;
 typedef unsigned char BYTE;
 
 typedef struct {
